+++ graal-nodejs/mx.graal-nodejs/com.oracle.truffle.trufflenode/src/com/oracle/truffle/trufflenode/GraalJSAccess.java
@@ -306,6 +306,7 @@ public final class GraalJSAccess {
     public static final TruffleString NODE_INTERNAL_GRAAL_WASM = Strings.constant("node:internal/graal/wasm");
     public static final TruffleString NODE_INTERNAL_WORKER_IO = Strings.constant("node:internal/worker/io");
     public static final TruffleString NODE_INTERNAL_MAIN_WORKER_THREAD = Strings.constant("node:internal/main/worker_thread");
+    public static final TruffleString NODE_GRAAL = Strings.constant("node:graal");
     public static final TruffleString NODE_INSPECTOR = Strings.constant("node:inspector");
     public static final TruffleString DOT_JS = Strings.constant(".js");
     public static final TruffleString SHEBANG = Strings.constant("#!");
@@ -2294,7 +2295,7 @@ public final class GraalJSAccess {
         }
     }
 
-    private static Object getExtraArgumentOfInternalScript(TruffleString moduleName, JSRealm realm) {
+    private Object getExtraArgumentOfInternalScript(TruffleString moduleName, JSRealm realm) {
         Object extraArgument = null;
         JSContext context = realm.getContext();
         if (NIO_BUFFER_MODULE_NAME.equals(moduleName)) {
@@ -2316,21 +2317,10 @@ public final class GraalJSAccess {
             extraArgument = (inspector == null) ? Null.instance : inspector;
         } else if (NODE_INTERNAL_GRAAL_WASM.equals(moduleName)) {
             extraArgument = createWasmStreamingCallback(realm);
+        } else if (NODE_GRAAL.equals(moduleName)) {
+            extraArgument = createEventLoopExecutor();
         }
         return extraArgument;
-    }
-
-    @TruffleBoundary
-    private static Object[] getInternalModuleUserArguments(Object[] args, TruffleString moduleName, JSRealm realm) {
-        Object[] userArgs = JSArguments.extractUserArguments(args);
-        Object extraArgument = getExtraArgumentOfInternalScript(moduleName, realm);
-        if (extraArgument == null) {
-            return userArgs;
-        }
-        Object[] extendedArgs = new Object[userArgs.length + 1];
-        System.arraycopy(userArgs, 0, extendedArgs, 0, userArgs.length);
-        extendedArgs[userArgs.length] = extraArgument;
-        return extendedArgs;
     }
 
     public Object scriptGetUnboundScript(Object script) {
@@ -3227,6 +3217,7 @@ public final class GraalJSAccess {
     }
 
     public void isolateDispose(boolean exit, int status) {
+        agent.setTaskRunnerPointer(0);
         if (exit) {
             exit(status);
         }
@@ -4179,6 +4170,19 @@ public final class GraalJSAccess {
     @TruffleBoundary
     private static void notifyWasmStreamingCallback(Object response, Object resolve, Object reject) {
         NativeAccess.notifyWasmStreamingCallback(response, resolve, reject);
+    }
+
+    private Object createEventLoopExecutor() {
+        return mainJSRealm.getEnv().asGuestValue(new EventLoopExecutor(agent));
+    }
+
+    public void isolateExecuteRunnable(Object runnable) {
+        try {
+            ((Runnable) runnable).run();
+        } catch (Exception ex) {
+            ex.printStackTrace();
+            System.exit(1);
+        }
     }
 
     public static class WeakCallback extends WeakReference<Object> {