+++ graal-nodejs/mx.graal-nodejs/com.oracle.truffle.trufflenode/src/com/oracle/truffle/trufflenode/NodeJSAgent.java
@@ -59,8 +59,12 @@ public class NodeJSAgent extends JSAgent {
         return thread;
     }
 
-    void setTaskRunnerPointer(long taskRunnerPointer) {
+    synchronized void setTaskRunnerPointer(long taskRunnerPointer) {
         this.taskRunnerPointer = taskRunnerPointer;
+    }
+
+    synchronized long getTaskRunnerPointer() {
+        return taskRunnerPointer;
     }
 
     @Override