+++ CHANGELOG.md
@@ -11,6 +11,7 @@ See [release calendar](https://www.graalvm.org/release-calendar/) for release da
 * Implemented the [Source Phase Imports](https://github.com/tc39/proposal-source-phase-imports) proposal. It is available behind the experimental option (`--js.source-phase-imports`).
 * Added option `js.stack-trace-api` that enables/disables `Error.captureStackTrace`, `Error.prepareStackTrace` and `Error.stackTraceLimit`. These non-standard extensions are disabled by default (unless `js.v8-compat` or `js.nashorn-compat` is used).
 * Made option `js.webassembly` stable.
+* Added an experimental `java.util.concurrent.Executor` that can be used to post tasks into the event loop thread in `graal-nodejs`. It is available as `require('node:graal').eventLoopExecutor`.
 
 ## Version 24.1.0
 * ECMAScript 2024 mode/features enabled by default.